refactor: reorganize component folders and fix imports
