+++ frontend/src/components/auth/ProtectedRoute.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
 import { useAuth } from '../../context/AuthContext';
-import Loading from '../common/Loading';
+import Loading from '../common/feedback/Loading';
 
 const ProtectedRoute = ({ children, requiredRole, allowedRoles }) => {
   const { isAuthenticated, user, isLoading } = useAuth();
