+++ frontend/src/components/auth/index.js
@@ -1,0 +1,11 @@
+// Authentication Components
+export { default as ProtectedRoute } from './ProtectedRoute';
+export { default as GuestOnlyRoute } from './GuestOnlyRoute';
+export { default as JobSeekerOnlyRoute } from './JobSeekerOnlyRoute';
+export { default as EmployerOnlyRoute } from './EmployerOnlyRoute';
+
+// Registration Steps
+export { default as PersonalInfoStep } from './PersonalInfoStep';
+export { default as AccountDetailsStep } from './AccountDetailsStep';
+export { default as SecurityStep } from './SecurityStep';
+export { default as StepProgress } from './StepProgress';
