+++ frontend/src/components/common/feedback/index.js
@@ -1,0 +1,4 @@
+// Feedback Components
+export { default as ErrorBoundary } from './ErrorBoundary';
+export { default as Loading } from './Loading';
+export { default as RoleAccessDenied } from './RoleAccessDenied';
