+++ frontend/src/components/common/forms/index.js
@@ -1,0 +1,4 @@
+export { default as FormInput } from './FormInput';
+export { default as FormSelect } from './FormSelect';
+export { default as FormTextarea } from './FormTextarea';
+// Add more form components as needed
