+++ frontend/src/components/dashboard/index.js
@@ -1,0 +1,20 @@
+// Shared Dashboard Components
+export { default as WelcomeSection } from './shared/WelcomeSection';
+export { default as DashboardSidebar } from './shared/DashboardSidebar';
+export { default as EmployerSidebar } from './shared/EmployerSidebar';
+export { default as StatsGrid } from './shared/StatsGrid';
+
+// Job Seeker Dashboard Components
+export { default as ProfileCompletionCard } from './jobseeker/ProfileCompletionCard';
+export { default as JobRecommendationsCard } from './jobseeker/JobRecommendationsCard';
+export { default as ProfileProgressCard } from './jobseeker/ProfileProgressCard';
+export { default as QuickActionsCard } from './jobseeker/QuickActionsCard';
+export { default as RecentActivityCard } from './jobseeker/RecentActivityCard';
+
+// Employer Dashboard Components
+export { default as WelcomeHeader } from './employer/WelcomeHeader';
+export { default as QuickActions } from './employer/QuickActions';
+export { default as RecentJobs } from './employer/RecentJobs';
+export { default as HiringPipeline } from './employer/HiringPipeline';
+export { default as UpcomingInterviews } from './employer/UpcomingInterviews';
+export { default as ActionAlert } from './employer/ActionAlert';
