+++ frontend/src/components/dashboard/jobseeker/ProfileCompletionCard.jsx
