+++ frontend/src/components/dashboard/jobseeker/ProfileProgressCard.jsx
