+++ frontend/src/components/dashboard/jobseeker/QuickActionsCard.jsx
