+++ frontend/src/components/dashboard/jobseeker/RecentActivityCard.jsx
