+++ frontend/src/components/dashboard/shared/EmployerSidebar.jsx
@@ -23,7 +23,7 @@ import {
   HiLogout,
 } from 'react-icons/hi';
 import { Link, useLocation } from 'react-router-dom';
-import { useAuth } from '../../context/AuthContext';
+import { useAuth } from '../../../context/AuthContext';
 
 const EmployerSidebar = () => {
   const location = useLocation();
