+++ frontend/src/components/dashboard/shared/WelcomeSection.jsx
