+++ frontend/src/components/index.js
@@ -1,0 +1,30 @@
+// Common Components
+export * from './common/feedback';
+export * from './common/forms';
+export * from './common/layout';
+export * from './common/ui';
+
+// Authentication Components
+export * from './auth';
+
+// Dashboard Components
+export * from './dashboard';
+
+// Profile Components
+export * from './profile';
+
+// Navigation Components
+export * from './navigation';
+
+// Job Components
+export * from './jobs';
+
+// Error Components
+export { default as ErrorActions } from './error/ErrorActions';
+export { default as ErrorBackground } from './error/ErrorBackground';
+export { default as ErrorContent } from './error/ErrorContent';
+export { default as ErrorHeader } from './error/ErrorHeader';
+export { default as FeaturesPreview } from './error/FeaturesPreview';
+
+// Section Components
+export * from './sections';
