+++ frontend/src/components/jobs/index.js
@@ -1,0 +1,6 @@
+// Job-related components (for future use)
+export { default as JobCard } from './JobCard';
+export { default as JobList } from './JobList';
+export { default as JobFilters } from './JobFilters';
+export { default as JobSearch } from './JobSearch';
+// Add job components as you create them
