+++ frontend/src/components/navigation/SideNavbar.jsx
