+++ frontend/src/components/navigation/index.js
@@ -1,0 +1,5 @@
+// Navigation Components
+export { default as SideNavbar } from './SideNavbar';
+
+// Dashboard Sidebars (these should probably be moved to dashboard/shared)
+// Will be updated after we organize dashboard components properly
