+++ frontend/src/components/profile/index.js
@@ -1,0 +1,14 @@
+// Job Seeker Profile Components
+export { default as AboutMe } from './jobseeker/AboutMe';
+export { default as Contact } from './jobseeker/Contact';
+export { default as Experience } from './jobseeker/Experience';
+export { default as Portfolio } from './jobseeker/Portfolio';
+export { default as ProfileHeader } from './jobseeker/ProfileHeader';
+export { default as ProfileQuickActions } from './jobseeker/ProfileQuickActions';
+export { default as Skills } from './jobseeker/Skills';
+
+// Employer Profile Components (when you create them)
+// export * from './employer';
+
+// Shared Profile Components
+// export * from './shared';
