+++ frontend/src/pages/jobseeker/JobSeekerProfile.jsx
@@ -23,13 +23,18 @@ import {
 } from '@chakra-ui/react';
 import { useAuth } from '../../context/AuthContext';
 
-import DashboardSidebar from '../../components/dashboard/DashboardSidebar';
-import ProfileHeader from '../../components/JobSeekerProfile/ProfileHeader';
-import AboutMe from '../../components/JobSeekerProfile/AboutMe';
-import Skills from '../../components/JobSeekerProfile/Skills';
-import Portfolio from '../../components/JobSeekerProfile/Portfolio';
-import Contact from '../../components/JobSeekerProfile/Contact';
-import Experience from '../../components/JobSeekerProfile/Experience';
+import {
+  ProfileHeader,
+  AboutMe,
+  Experience,
+  Skills,
+  Portfolio,
+  Contact,
+  ProfileQuickActions
+} from '../../components/profile';
+
+import { DashboardSidebar } from '../../components/dashboard';
+import { Loading } from '../../components/common/feedback';
 
 const JobSeekerProfile = () => {
   const [activeTab, setActiveTab] = useState(0);
